move database to library package
