+++ backend/database/database.go
@@ -1,4 +1,4 @@
-package main
+package database
 
 import (
 	"crypto/sha256"
